refactor: move functions to set up model state for tests to a test utils file
